+++ backend/README.md
@@ -24,3 +24,15 @@
 - [NestJS와 ioredis를 이용한 초대링크 만료 시간 설정하기](https://threeyears.tistory.com/599)
 - [NestJS에서 Redis를 활용한 최근 검색어 기능 구현](https://threeyears.tistory.com/614)
 - [NestJS에서 Redis를 활용한 좋아요 성능 개선](https://threeyears.tistory.com/621)
+
+# 그룹 간의 팔로우 기능
+
+1. A 사용자가 A그룹에 가입, B그룹에 가입 C그룹에 가입
+2. B 사용자가 D그룹에 가입, E그룹에 가입, F그룹에 가입
+3. A 그룹과 D그룹이 팔로우, D그룹이 A그룹을 팔로우 (양방향 팔로우)
+4. A 그룹이 E그룹에 팔로우, E그룹이 A그룹을 팔로우 (양방향 팔로우)
+5. F그룹이 A그룹에 팔로우 (단방향 팔로우)
+6. A 사용자가 A그룹에 피드를 작성 -> 양방향 팔로우 관계인 D그룹과 E그룹의 멤버들한테 피드가 노출.
+7. F그룹은 단방향 팔로우 관계이기 때문에 F그룹에는 멤버들에게는 해당 피드가 노출되지 않음.
+8. 근데 만약 A 사용자가 A그룹에 피드를 작성 할때 isVisibleToFollowers 속성을 true로 설정 했다면?
+9. 이럴경우 F그룹의 멤버들에게 피드가 노출되고 + 양방향 팔로우인 D그룹과 E그룹의 멤버들에게도 피드가 노출됨.
