+++ backend/src/api/groups/groups.service.ts
@@ -1,8 +1,7 @@
-import { Injectable, OnModuleInit } from '@nestjs/common';
+import { Injectable } from '@nestjs/common';
 import { Not, QueryRunner } from 'typeorm';
 import { v4 as uuidv4 } from 'uuid';
 
-import { QueryRunnerWithRedis } from '@/common/decorators/query-runner-with-redis.decorator';
 import {
 	EntityConflictException,
 	EntityNotFoundException,
@@ -17,13 +16,11 @@ import {
 	ERROR_NO_PERMISSTION_TO_GROUP,
 } from '@/constants/business-error';
 import { MAIN_ROLE } from '@/constants/string-constants';
-import { GroupFollowCache } from '@/models/cache/group-follow-cache';
 import { BelongToGroupResDto } from '@/models/dto/group/res/belong-to-group.res.dto';
 import { GroupMembersResDto } from '@/models/dto/group/res/group-members.res.dto';
 import { GroupProfileResDto } from '@/models/dto/group/res/group-profile.rest.dto';
 import { GroupResDto } from '@/models/dto/group/res/group-res.dto';
 import { FamsRepository } from '@/models/repositories/fams.repository';
-import { GroupFollowRepository } from '@/models/repositories/group-follow.repository';
 import { GroupsRepository } from '@/models/repositories/groups.repository';
 import {
 	ICreateGroupArgs,
@@ -34,84 +31,11 @@ import {
 import { getOffset } from '@/utils/getOffset';
 
 @Injectable()
-export class GroupsService implements OnModuleInit {
+export class GroupsService {
 	constructor(
 		private readonly groupsRepository: GroupsRepository,
 		private readonly famsRepository: FamsRepository,
-		private readonly groupFollowCache: GroupFollowCache,
-		private readonly groupFollowRepository: GroupFollowRepository,
 	) {}
-
-	/**
-	 * 서버 시작 시 모든 그룹의 팔로우를 Redis와 동기화 (Cache Warming)
-	 */
-	async onModuleInit() {
-		const groupIds = await this.groupsRepository.findAllGroupIds();
-		for (const groupId of groupIds) {
-			const [followers, following] = await Promise.all([
-				this.groupFollowRepository.getFollowers(groupId),
-				this.groupFollowRepository.getFollowings(groupId),
-			]);
-
-			await this.groupFollowCache.syncFollows(
-				groupId,
-				followers.map((data) => data.followingGroupId),
-				following.map((data) => data.followedGroupId),
-			);
-		}
-	}
-
-	/**
-	 * 레디스에서 해당 그룹을 팔로우하는 그룹들을 가져온다.
-	 */
-	async getFollowers(groupId: string) {
-		return await this.groupFollowCache.getFollowers(groupId);
-	}
-
-	/**
-	 * 레디스에서 해당 그룹을 팔로우하는 그룹들을 가져온다.
-	 */
-	async getFollowings(groupId: string) {
-		return await this.groupFollowCache.getFollowing(groupId);
-	}
-
-	async followGroup(
-		followingGroupId: string,
-		followedGroupId: string,
-		qrAndRedis: QueryRunnerWithRedis,
-	) {
-		const { queryRunner, redisMulti } = qrAndRedis;
-		const isFollowing = await this.groupFollowCache.isFollowing(
-			followingGroupId,
-			followedGroupId,
-		);
-
-		if (isFollowing) {
-			await this.groupFollowCache.removeFollower(
-				followedGroupId,
-				followingGroupId,
-				redisMulti,
-			);
-			await this.groupFollowRepository.removeFollow(
-				followingGroupId,
-				followedGroupId,
-				queryRunner,
-			);
-		} else {
-			await this.groupFollowCache.addFollower(
-				followedGroupId,
-				followingGroupId,
-				redisMulti,
-			);
-			await this.groupFollowRepository.addFollow(
-				followingGroupId,
-				followedGroupId,
-				queryRunner,
-			);
-		}
-
-		return !isFollowing;
-	}
 
 	async getMemberListBelongToGroup({
 		groupId,
